requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.17.0
xgboost==1.5.2
python-dotenv==1.0.0
kaggle==1.5.16
streamlit-lottie==0.0.5
